Consolidate tracing: make provider-agnostic and merge distributed utilities
- Replace DataDog-specific environment variables with OpenTelemetry standard variables
- Follow op-node patterns for consistent OP Stack initialization
- Use OTEL SDK defaults for automatic configuration via standard env vars
- Merge distributed.go functionality into tracer.go for single-file tracing package
- Update InitializeTracing() to require serviceName parameter
- Remove manual OTLP endpoint parsing - let SDK handle configuration
- Fix linter issues: use 'any' instead of 'interface{}', format all files

Key environment variables now supported:
- OTEL_SERVICE_NAME (primary) with DD_SERVICE fallback
- OTEL_EXPORTER_OTLP_ENDPOINT for any OTLP-compatible backend
- OTEL_RESOURCE_ATTRIBUTES for additional metadata
